+++ packages/db/src/index.ts
@@ -23,4 +23,5 @@ export * from "./repositories/postgres/workflowExecutionRepository.js";
 export * from "./repositories/postgres/taskExecutionRepository.js";
 export * from "./repositories/postgres/executionEventRepository.js";
 export * from "./repositories/postgres/deadLetterRepository.js";
+export * from "./repositories/postgres/memoryRecordRepository.js";
 export * from "./repositories/memory/inMemoryRepositories.js";